Support PipelineBehavior Beans
Previously pipeline behaviors needed to be registered manually and
jmediator and execute those references as part of the pipeline. As a
result, we did not support pipelines behavior that were registered as
beans via various frameworks classpath scanning.

This changes replaces RequestHandlerProvider with a ServiceFactory. The
implementation for requests handlers is essentially unchanged between
the two however ServiceFactory exposes a way to get pipeline behaviors.
This allows the various implementations to resolve behaviors as needed
including resolving via IoC frameworks.