Quarkus Extension
This adds a Jmediator Quarkus extension which is broken out into two
modules deployment and runtime.

Deployment: Handles buildtime augmentation by specifying a series of
build steps to apply to java source files and includes compilation, annotation processing, bytecode generation.
The output is recorded bytecode which is directly responsible for
directly instantiating runtime services.
This is essentially finding all request handlers and pipeline behaviors
on the classpath and registering them with the ServiceFactory.

Runtime: Work that is executed when you run the application.
Represents the capabilities the extension developer exposes to the application’s developer.
In the case of Jmediator that is the ServiceFactory and RequestDispatcher.

There is still a bit of clean up that needs to occur. I have a lot of
junk still sitting around that I need to delete. I have have a small bug
where quarkus is removing requesthandler and pipeline behavior beans
because they are not direclty referenced anywhere.