Pass request via PipelineChain doBehavior
Exploring a scenario in which I wanted to build a pipeline behavior that alters
the request object and pass new object down the chain.

Specifically I wanted to add additiona behavior by wrapping the request
object in another request decorator.

This commit enables this scnario by updating the PipelineChain deBehavior
method to accept a request objects.