Support Changing Request Types Within Pipeline
This commit allows pipeline behaviors to alter the request type during
the pipeline. Use case is to allow pipelines to decorate requests as
part of the pipeline. A concrete example might be that as part of your
pipeline you want to add current user information. This allows you to
wrap the incoming request in a request that allows you to provide
current user and pass that down the pipeline and eventually to the
handler